Refactoring, renamed manual test.
