--- tests/manual/imageupload.md
@@ -1,10 +1,0 @@
-## Image upload
-
-1. Drop an image into editor.
-1. Image should be read and displayed.
-1. Press "Upload progress" button couple times to simulate upload process.
-1. After uploading is complete your image should be replaced with sample image from server.
-
-Repeat all the steps with:
-* dropping multiple images,
-* using toolbar button to add one and multiple images.
