+++ tests/manual/uploadadapter.html
@@ -2,16 +2,3 @@
 	<h2>Image upload</h2>
 	<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla finibus consequat placerat. Vestibulum id tellus et mauris sagittis tincidunt quis id mauris. Curabitur consectetur lectus sit amet tellus mattis, non lobortis leo interdum.</p>
 </div>
-
-<div id="button-container"></div>
-
-<style>
-	#button-container div {
-		margin-top: 10px;
-	}
-
-	#button-container button {
-		margin-right: 10px;
-	}
-</style>
-
