+++ tests/manual/uploadadapter.js
@@ -21,14 +21,17 @@ import Italic from '@ckeditor/ckeditor5-basic-styles/src/italic';
 import List from '@ckeditor/ckeditor5-list/src/list';
 import ImageUpload from '@ckeditor/ckeditor5-upload/src/imageupload';
 import ImageUploadProgress from '@ckeditor/ckeditor5-upload/src/imageuploadprogress';
-import CKFinderAdapter from '../../src/ckfinderadapter';
+import CKFinderUploadAdapter from '../../src/uploadadapter';
 
 ClassicEditor.create( document.querySelector( '#editor' ), {
 	plugins: [
 		Enter, Typing, Paragraph, Heading, Undo, Bold, Italic, Heading, List, Image, ImageToolbar, Clipboard,
-		ImageCaption, ImageStyle, ImageUpload, CKFinderAdapter, ImageUploadProgress
+		ImageCaption, ImageStyle, ImageUpload, CKFinderUploadAdapter, ImageUploadProgress
 	],
-	toolbar: [ 'headings', 'undo', 'redo', 'bold', 'italic', 'bulletedList', 'numberedList', 'insertImage' ]
+	toolbar: [ 'headings', 'undo', 'redo', 'bold', 'italic', 'bulletedList', 'numberedList', 'insertImage' ],
+	ckfinder: {
+		uploadUrl: 'https://cksource.com/weuy2g4ryt278ywiue/core/connector/php/connector.php?command=QuickUpload&type=Files&responseType=json'
+	}
 } )
 .catch( err => {
 	console.error( err.stack );
